add leetcode-23's test.
</subject>

Branch:

<type>:
- [ ] Bug fix
- [ ] Bug fix (Test)
- [x] New feature
- [ ] Breaking change
- [ ] Documentation update
- [ ] This change requires a documentation update

<body>

<footer>

+++ cpp/source/leetcode_20_.hpp
@@ -48,7 +48,7 @@ class Solution20 {
 public:
     bool isValid(const string &s) {
         std::stack<uint8_t> stk;
-        std::unordered_map<int32_t, int32_t> umap;
+        std::unordered_map<uint8_t, uint8_t> umap;
         umap['('] = ')';
         umap['['] = ']';
         umap['{'] = '}';
